pyproject.toml
[project]
name = "trademind"
version = "0.1.0"
description = "Blockchain anomaly detection and whale tracking system"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "pytest>=7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/trademind"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
